rename components to component
The other directories use singularity.
