+++ javascript/experiments/maze/src/component/Canvas.js
