+++ javascript/experiments/maze/src/component/Maze/algorithm/createInnerEdges.js
