+++ javascript/experiments/maze/src/component/Maze/algorithm/createMazePath.js
