+++ javascript/experiments/maze/src/component/Maze/class/Cell.js
