+++ javascript/experiments/maze/src/component/Maze/class/Edge.js
